change a lot of I8 to Int
the reason for these changes is that
std:int8_t is just typedef'd to signed char,
which the C/C++ does not always interpret
as an integer, because it has special
meaning as a character.
the main issue right now is every time
we try to print an I8 we get a strange
character to screen.
Only the array storage type and the
type of alignment codes needs to be I8.