copy_meta helper function for modification
this copies all the metadata that would
usually need to be copied when:
1) migrating
2) reordering
3) adapting
it is important because these are seven
or so lines of code that are copy pasted
in all the above use cases, and when
another piece of metadata is added there
should only be one place to modify.
we have records from omega_h of bugs
arising from this, where new metadata
was added and some adapt functions preserved
it but migration forgot to, etc.