fix binary file adjacencies
also put in place checks at
runtime and compile time that
would have prevented the mistake.
at runtime, add_adj() checks that
the fields of Adj that are needed
do indeed exist.
at compile time, there was an implicit
conversion being made between LOs and Adj,
which is highly misleading.
Mark the appropriate constructors as
explicit to avoid this confusion.