go ahead and canonicalize all derived entities
I had thought that we could skip exposed boundary
sides, but there is a contradication in that we
cannot easily determine which sides are exposed
in parallel without first having parallel consistency
and ownership, so we cannot use exposure information
when making things consistent to begin with.
It seems wholly acceptable to have inward-curling
sides when the user gives us a parallel mesh that
only specifies element connectivity.
If we want to keep them outward-curling, it will
have to be another step after their initial consitency
is established.