don't repeat path name when writing VTK
we used to have something like:

./gold/gold.pvd
./gold/steps/step_0/...

now the .pvd file is also statically named, i.e.

./gold/steps.pvd

the original reason for repeating the paths is
that in Paraview's pipeline viewer that is the
name of the mesh (steps.pvd), so it would be
hard to distinguish different entries.
however, giving them unique names makes it hard
to move the data set, i.e.

mv ./result/ ./gold/

because it would not find "./gold/result.pvd",
and it would be misleading anyway.
for now we go with all static names.