add interpolate_metrics helper
due to our quality-sensitive adapt system,
and the fact that anisotropic requests will
often make the input mesh look low-quality
in metric space to begin with, I think it
makes sense to have a similar approach to
warp_to_limit() where we gradually increase
the anistropy over several calls to adapt().
this helper is the first step to doing that.