allowable generated quality is floor, not ceiling
There are two quality values passes into adaptation,
the floor and ceiling.
The idea at a higher level (taking into account the
warp_to_limit() system) is that there is an element
quality we are willing to temporarily tolerate
but never go below (the floor) and another that we
would like to satisfy on the output, "final" mesh.
warping would decrease quality into this buffer zone,
and adaptation would raise it back up.
Until now, the qual_floor argument only affected
the debug output of the adapt() call, not its
behavior.
This change allows refinement and coarsening
to create elements with quality lower than
qual_ceil but higher than qual_floor.
The idea being that we may want to allow refinement
to create lower-than-ideal quality elements
which are subsequently repaired by quality improvement.
The old behavior can be preserved by setting
qual_floor=qual_ceil when calling adapt(),
which we've done here for the tests that already
run successfully.