add shared library option with full RPATH
The RPATH configuration is such that
the executables and libraries get all the
RPATH entries they need to find their dependencies
(It is all geared towards the install directory).
This is useful if you install several versions
of Omega_h or its dependencies, for example
a version compiled with GCC and another with Clang,
or a CUDA-oriented version and a pure serial version.
With adequate control of installation directories,
each executable/library will link to its appropriate
dependencies, alleviating the need for LD_LIBRARY_PATH
and other environment variable help.