README: no standalone Kokkos
We now rely more on the CMake
packaging of Kokkos, which means
the standalone Makefile build
of Kokkos is no longer something we can
link to.
However, there is ongoing work to give
Kokkos a standalone CMake build, and if
that has decent exported targets then we
can recommend that as well
(see kokkos/kokkos#155).