size field scaling for nelems, new shape math
this work includes a system to scale a given
isotropic or anisotropic size field such
that adapting to the scaled size field
produces roughly a certain number of elements.

also motivated by this, the shape functions
were rewritten to more intuitive and hopefully
faster forms.
however, this is not bitwise backwards compatible,
so this is going in a branch at the moment.