use project() for identity size field code
this should be much more accurate in both
cases.
in the isotropic case, for simple uniform
meshes we no longer see different answers
at the boundary due to lack of adjacent nodes.
we now use the root mean squared edge length
of an element as its isotropic size field value,
then project to nodes.
anisotropic metrics are "linearized" and then
projected, which is untested but should be
more accurate because the projection method
switch from volume average to boundary-correcting
SPR.