first draft QR overhaul
1) try to allow runtime-sized problems
with compile-time sized data structures.
this will support solving linear systems
where the size of the system depends on the
number of elements in a cavity and we have
to preallocate some maximum system size
but we don't need to waste time solving
the maximum size.

2) remove traces of QR-based eigensolver.
no matter what we tried, that solver couldn't
handle non-Hermitian matrices so it was removed earlier.

3) have a QRFactorization class, use that
in the fit.hpp code. fit.hpp was also updated
to use the runtime problem sizing feature.

Right now this compiles but has not been run.