slight change to metric limiting
the average edge length is estimated
using the two metrics pre-limiting,
and if the metric to be limited was
asking for long edges then very little
variance will be accepted and a strong
limit will be imposed.
Instead, use the "other" metric as the
sole indicator of edge length and thus rate.