limit approach by edge length as well
the initial UGAWG cube mesh is too coarse
to resolve the given metric using our
quality-based approach alone.
I got it to work by also limiting the
approach based on edges being twice their
maximum desired length.
I'm not sure how viable it is to keep this
in the long run, but for now it works on
this case and the uniform aniso test I have...