better cache invalidation
there are only three hard things...

if we read a .osh file that has the
length field on it, then set a new
size field, the previous logic would
not invalidate the length field read
from file.
this new logic adds a boolean flag
to explicitly mark those set_tag
calls which are special internal
calls guaranteed not to need
invalidation.