initialize parting_ and nghost_layers_ to -1
this seems to have been broken a long time
ago: these variables are supposed to start
out as -1 for uninitialized so things like
set_parting can detect the uninitialized case
and just accept values without doing anything.
maybe we could clean this up like we did
for cache invalidation, for now this is a
step in the right direction.