safety check and barrier in binary IO
if one tries to read a parallel directory
right after writing it, rank 0 may not be
done filling in the nparts file before
another rank tries to read it.
the barrier ensures that the write function
does not return until the nparts file is closed.
also ensure that bad parsing of the nparts file properly
triggerss an error instead of returning an
uninit value.