improve Few copy ctor
it was using assignment on the
array elements before, but when
T=Write<T2> that isn't good enough,
it needs to call the copy constructor.
I used placement-new to do this.

There may be an issue with overwriting
an object that has already been
constructed... so far, not an issue though.