trying to better deal with new interior subcavs
part of (#106)
there can be swaps and refinements such that
before the operation, all elements in the
cavity were touching the boundary, and after
the operation there are elements in the
cavity that don't touch the boundary.
I handn't initially considered this case,
now I'm trying to get the code to output
zero integral errors for those new interior
elements.