compute integral error on full cavity (#106)
conservation is not guaranteed at a sub-cavity
level; only at the full-cavity level (and
even then only in certain cases).
Thus we have to lift the computatio of
newly introduced error out of the subcavity
code, and re-express things as densities
to make the transition to the sub-cavity
level easier.