got roped into more refactoring ! (#106)
the 3d_conserve_test was showing no mass
errors in the object that was shrinking.
this is due to mass error not being
computed on a per-material basis.
I needed to get more organized with
cavity sets, so divided them by
two dimensions of three categories each.
also, density was getting low near the
boundary on the object that was gaining
volume. I think density remap at boundary
collapses should just copy from the
old elements, since the idea is that remap
should preserve the distribution of
density as much as possible.