downgrade raise() due to CUDA complaints
apparently the CUDA compiling wasn't
expanding the recursive constexpr calls
into a single expression, and thus
was compilaining that it couldn't determine
stack size for a kernel that uses raise()
due to the recursion.
raise() is now implemented using template
recursion, which the CUDA compiler seems
to accept and will probably be more reliably
optimizable.