fix critical bug in length histogram computation
the edge length data was being treated as vertex
data, meaning that many data points were ignored.
the only reason this didn't segfault is because
in 2D and 3D there are many more edges than vertices.
it took the 1D test to inform us of the issue !
(there is one more vertex than edges in 1D)
I've added checks in the histogram and owned_array
code that will trip if this kind of mistake is made
again, and I added histogram printing to the
warp_test so that histograms get tested somewhere
besides the 1D tests.