Prepare for face based variables during mesh changes.
There are physics that have face based variables
that will have to be handled during mesh topology
changes to preserve certain attributes.  How this
will be done will depend on the physics.  This
small push simply allows for the existence of
such variables and only provides a placeholder
of where to put the calculation of new face
values from the old once the correct algorithm
is determined.