Fix test with one proc / add tests with 2 procs
@ibaned: please note the ugly copy of
`copies2data` in `create_dist_for_variable_sized`
I wish it was possible to create a osh::Read over a subset of a Write,
in this case [0, size(Write) -2].

We could add a `Int LengthCap = 0` template parameter to `osh::Read`,
and more or less use it this way:
```cpp
template <typename T, Int LengthCap>
class Read {
 public:
    OMEGA_H_INLINE LO size() const OMEGA_H_NOEXCEPT { return write_.size() - LengthCap; }
```
It really don't like this solution as it would require ugly explicit
template instantiations for LengthCap equal to 0, 1, ...
But it would not introduce a memory nor a runtime cost with the default case.

Do you see a nicer solution to avoid this copy?