Fixup changes from cherry-pick
The original attempt at mergeing sandialabs omega_h failed to bring over
the git commit history. However, I had made some important changes to
the rc_fields on top of that merge. To bring over the new changes I
cherry-picked the commits, but there were a few mistakes that came with
the chery-pick results. This commit fixes up all of those issues so that
the build compiles and all tests pass.